feat(): :construction: improve接口改造

+++ src/services/screen.ts
@@ -1,13 +1,49 @@
-import { isModelHash, useAnyDva } from '@/hooks';
 import { merge } from 'lodash';
+import { isModelHash, useAnyDva } from '@/hooks';
+import GlobalConfig from '@/utils/Assist/GlobalConfig';
 import { getLocationQuery } from '../utils';
 import { SCREEN_VERSION, SERVICE_REQUEST_URL } from '../utils/constants';
 import request from '../utils/request';
+import {
+  postScreen4Improve,
+  putScreen4Improve,
+  putScreenPool4Improve,
+  createPutScreenPool4Improve,
+  putScreenPoolValid4Improve,
+  deleteScreenPool4Improve,
+  getScreenDetail4Improve,
+  getScreenList4Improve,
+  deleteScreen4Improve,
+  previewScreen4Improve,
+  previewScreenValid4Improve,
+  copyScreen4Improve,
+  shareScreen4Improve,
+  closeShareScreen4Improve,
+  shareScreenHeartbeat4Improve,
+  shareScreenGet4Improve,
+  shareScreenPost4Improve,
+  enableScreen4Improve,
+  disabledScreen4Improve,
+  preRequestData4Improve,
+  postScreenLeadIn4Improve,
+  postScreenExport4Improve,
+} from './improve';
 
 const { REACT_APP_ENV } = process.env;
 
+export {
+  getScreenShotList,
+  getCurrentScreenShotData,
+  deleteScreenShot,
+  updateScreenShot,
+  addScreenShot,
+  useScreenShot,
+  coverScreenShot,
+} from './improve';
+
 // 新增大屏
 export const postScreen = (data: API_SCREEN.TAddScreenParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return postScreen4Improve(data);
   return request('/api/screen/list', {
     method: 'POST',
     data: {
@@ -19,6 +55,7 @@ export const postScreen = (data: API_SCREEN.TAddScreenParams) => {
 
 // 修改大屏
 export const putScreen = (data: API_SCREEN.TEditScreenParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return putScreen4Improve(data);
   return request('/api/screen/list', {
     method: 'PUT',
     data: {
@@ -30,6 +67,7 @@ export const putScreen = (data: API_SCREEN.TEditScreenParams) => {
 
 // 链式修改大屏
 export const putScreenPool = (data: API_SCREEN.TEditScreenPoolParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return putScreenPool4Improve(data);
   return request('/api/screen/list/pool', {
     method: 'PUT',
     data: {
@@ -43,6 +81,7 @@ export const putScreenPool = (data: API_SCREEN.TEditScreenPoolParams) => {
 export const createPutScreenPool = (
   data: API_SCREEN.TCreateScreenPoolParams,
 ) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return createPutScreenPool4Improve(data);
   return request('/api/screen/list/pool', {
     method: 'POST',
     data,
@@ -51,6 +90,8 @@ export const createPutScreenPool = (
 
 // 大屏编辑状态验证
 export const putScreenPoolValid = (params: { _id: string }) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND)
+    return putScreenPoolValid4Improve(params);
   return request('/api/screen/list/pool', {
     method: 'GET',
     params,
@@ -62,6 +103,8 @@ export const deleteScreenPool = (
   sync: boolean = false,
   customParams?: Partial<{ _id: string; type: string }>,
 ) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND)
+    return deleteScreenPool4Improve(sync, customParams);
   const { id } = getLocationQuery() || {};
   const { getState } = useAnyDva();
   const userId = getState().user.currentUser._id;
@@ -88,6 +131,7 @@ export const deleteScreenPool = (
 
 // 大屏详情
 export const getScreenDetail = (params: API_SCREEN.TGetScreenDetail) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return getScreenDetail4Improve(params);
   return request<API_SCREEN.TScreenDetail>('/api/screen/detail', {
     method: 'GET',
     params,
@@ -96,6 +140,7 @@ export const getScreenDetail = (params: API_SCREEN.TGetScreenDetail) => {
 
 // 大屏列表
 export const getScreenList = (params: API_SCREEN.TGetScreenListParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return getScreenList4Improve(params);
   return request<any>('/api/screen/list', {
     method: 'GET',
     params,
@@ -105,6 +150,7 @@ export const getScreenList = (params: API_SCREEN.TGetScreenListParams) => {
 
 // 大屏删除
 export const deleteScreen = (params: API_SCREEN.TDeleteScreenParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return deleteScreen4Improve(params);
   return request('/api/screen/list', {
     method: 'DELETE',
     params,
@@ -113,6 +159,7 @@ export const deleteScreen = (params: API_SCREEN.TDeleteScreenParams) => {
 
 // 大屏预览
 export const previewScreen = (data: API_SCREEN.TPreviewScreenParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return previewScreen4Improve(data);
   return request('/api/screen/preview', {
     method: 'POST',
     data,
@@ -121,6 +168,8 @@ export const previewScreen = (data: API_SCREEN.TPreviewScreenParams) => {
 
 // 大屏预览验证
 export const previewScreenValid = (params: API_SCREEN.TPreviewScreenParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND)
+    return previewScreenValid4Improve(params);
   return request('/api/screen/preview/valid', {
     method: 'GET',
     params,
@@ -129,6 +178,7 @@ export const previewScreenValid = (params: API_SCREEN.TPreviewScreenParams) => {
 
 // 大屏复制 | 大屏模板使用
 export const copyScreen = (data: { _id: string; type: 'screen' | 'model' }) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return copyScreen4Improve(data);
   return request('/api/screen/copy', {
     method: 'POST',
     data,
@@ -140,6 +190,7 @@ export const copyScreen = (data: { _id: string; type: 'screen' | 'model' }) => {
 // 设置公共或加密
 // 设置有效时间
 export const shareScreen = (data: API_SCREEN.TShareScreenParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return shareScreen4Improve(data);
   return request('/api/screen/share', {
     method: 'POST',
     data,
@@ -150,6 +201,7 @@ export const shareScreen = (data: API_SCREEN.TShareScreenParams) => {
 export const closeShareScreen = (
   params: API_SCREEN.TCloseShareScreenParams,
 ) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return closeShareScreen4Improve(params);
   return request('/api/screen/share', {
     method: 'DELETE',
     params,
@@ -160,6 +212,8 @@ export const closeShareScreen = (
 export const shareScreenHeartbeat = (
   params: API_SCREEN.TShareScreenHeartbeatParams,
 ) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND)
+    return shareScreenHeartbeat4Improve(params);
   return request('/api/screen/share', {
     method: 'GET',
     params,
@@ -168,6 +222,7 @@ export const shareScreenHeartbeat = (
 
 // 大屏分享权限信息获取
 export const shareScreenGet = (params: API_SCREEN.TShareScreenGetParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return shareScreenGet4Improve(params);
   return request<API_SCREEN.TShareScreenGetData>('/api/screen/share/valid', {
     method: 'GET',
     params,
@@ -177,6 +232,7 @@ export const shareScreenGet = (params: API_SCREEN.TShareScreenGetParams) => {
 
 // 大屏分享权限验证
 export const shareScreenPost = (data: API_SCREEN.TShareScreenPostParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return shareScreenPost4Improve(data);
   return request('/api/screen/share/valid', {
     method: 'POST',
     data: {
@@ -188,6 +244,7 @@ export const shareScreenPost = (data: API_SCREEN.TShareScreenPostParams) => {
 
 // 大屏启用
 export const enableScreen = (data: API_SCREEN.TEnableScreenParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return enableScreen4Improve(data);
   return request('/api/screen/enable', {
     method: 'PUT',
     data,
@@ -196,6 +253,7 @@ export const enableScreen = (data: API_SCREEN.TEnableScreenParams) => {
 
 // 大屏禁用
 export const disabledScreen = (params: API_SCREEN.TDisabledScreenParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return disabledScreen4Improve(params);
   return request('/api/screen/enable', {
     method: 'DELETE',
     params,
@@ -204,6 +262,7 @@ export const disabledScreen = (params: API_SCREEN.TDisabledScreenParams) => {
 
 // 服务端代理数据请求
 export const preRequestData = (data: API_SCREEN.TPreRequestDataParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return preRequestData4Improve(data);
   return request(SERVICE_REQUEST_URL, {
     method: 'POST',
     data,
@@ -212,6 +271,7 @@ export const preRequestData = (data: API_SCREEN.TPreRequestDataParams) => {
 
 // 大屏导入
 export const postScreenLeadIn = (data: API_SCREEN.TPreLeadInDataParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return postScreenLeadIn4Improve(data);
   return request('/api/screen/pre/leadin', {
     method: 'POST',
     data,
@@ -220,6 +280,7 @@ export const postScreenLeadIn = (data: API_SCREEN.TPreLeadInDataParams) => {
 
 // 大屏导出
 export const postScreenExport = (data: API_SCREEN.TPreExportDataParams) => {
+  if (GlobalConfig.IS_IMPROVE_BACKEND) return postScreenExport4Improve(data);
   return request('/api/screen/pre/export', {
     method: 'POST',
     data,
